TW-2359: refactoring
