+++ src/screens/manage-accounts/hooks/use-filtered-account-list.ts
@@ -1,9 +1,9 @@
 import { useMemo, useState } from 'react';
 
-import { Account } from 'src/interfaces/account.interfaces';
-import { useGetSaplingAddressForAccount } from 'src/store/sapling/sapling-selectors';
+import { Account } from 'src/interfaces/account.interfaces.ts';
+import { useGetSaplingAddressForAccount } from 'src/store/sapling/sapling-selectors.ts';
 import { accountMatchesSearch } from 'src/utils/account.utils.ts';
-import { isString } from 'src/utils/is-string';
+import { isString } from 'src/utils/is-string.ts';
 
 export const useFilteredAccountList = (accountList: Account[]) => {
   const [searchValue, setSearchValue] = useState<string>();
