+++ src/screens/manage-accounts/manage-hd-accounts/manage-account-item/selectors.ts
