+++ src/screens/manage-accounts/manage-hd-accounts/manage-account-item/styles.ts
