+++ src/screens/manage-accounts/manage-hd-accounts/styles.ts
